requirements.txt
django
djangorestframework
pygments
coreapi
coreapi-cli
